Closes #184.

Perhaps, the users of `big` may need more accuracy than speed.(cf. https://github.com/JuliaMath/FixedPointNumbers.jl/issues/129#issuecomment-568863687)
I want to stop using the reciprocal technique below: (**Edit:** done.)
https://github.com/JuliaMath/FixedPointNumbers.jl/blob/7c90fb6b48867ff2fda1608b0b9c5b140e0ad963/src/normed.jl#L247
~For large arrays, using the reciprocal technique seems to allocate more memory and slow.~
In the first place, it seems that the constant folding is not applied. This seems to be essentially different from the following `Float64`'s case.
https://github.com/JuliaMath/FixedPointNumbers.jl/blob/7c90fb6b48867ff2fda1608b0b9c5b140e0ad963/src/FixedPointNumbers.jl#L99-L100