@kimikage 2020-07-11
```julia
julia> versioninfo()
Julia Version 1.4.2
Commit 44fa15b150* (2020-05-23 18:35 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-8.0.1 (ORCJIT, skylake)
```

Before:
```julia
julia> @benchmark BigFloat.(x) setup=(x=collect(rand(N0f8, 1000, 1000)))
BenchmarkTools.Trial:
  memory estimate:  328.06 MiB
  allocs estimate:  6000002
  --------------
  minimum time:     508.701 ms (38.52% GC)
  median time:      589.345 ms (37.99% GC)
  mean time:        631.865 ms (34.44% GC)
  maximum time:     829.537 ms (43.97% GC)
  --------------
  samples:          9
  evals/sample:     1

julia> BigFloat(5N32f32)
5.000000000000000000000000000000000000000000000000000000000000000000000000000069
```
After:
```julia
julia> @benchmark BigFloat.(x) setup=(x=collect(rand(N0f8, 1000, 1000)))
BenchmarkTools.Trial:
  memory estimate:  221.25 MiB
  allocs estimate:  4000002
  --------------
  minimum time:     325.089 ms (28.13% GC)
  median time:      450.361 ms (29.96% GC)
  mean time:        442.812 ms (31.00% GC)
  maximum time:     536.980 ms (40.39% GC)
  --------------
  samples:          12
  evals/sample:     1

julia> big(5N32f32)
5.0
```