@kimikage 2020-07-12
I think there is room for improvement in the handling of `BigFloat`, but we should not need to improve it in this PR.